pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-input"
version = "1.0.1"
description = "Windows tray application for voice input with cloud-based (Google Gemini) and local model transcription services"
readme = "README.md"
authors = [{name = "ErlichLiu", email = "erlichliu@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "cachetools==5.5.2",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fastapi==0.116.1",
    "fonttools==4.59.0",
    "google-auth==2.40.3",
    "google-genai==1.29.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.5",
    "numpy==2.2.1",
    "opencc-python-reimplemented==0.1.7",
    "packaging==24.2",
    "pillow==11.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.10.5",
    "pydantic-core==2.27.2",
    "pynput==1.7.7",
    "pyparsing==3.2.3",
    "pyperclip==1.9.0",
    "pystray==0.19.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "requests==2.32.3",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sounddevice==0.5.1",
    "soundfile==0.13.0",
    "starlette==0.47.2",
    "tenacity==9.1.2",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
    "pycaw",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.0",
    "pip-tools>=7.4.0",
]

[project.scripts]
whisper-input = "src.windows_app:main"

[project.gui-scripts]
whisper-input-gui = "src.windows_app:main"

[project.urls]
Homepage = "https://github.com/ErlichLiu/Whisper-Input"
Repository = "https://github.com/ErlichLiu/Whisper-Input"
Issues = "https://github.com/ErlichLiu/Whisper-Input/issues"

[tool.setuptools]
packages = ["src", "src.audio", "src.keyboard", "src.llm", "src.transcription", "src.utils"]
py-modules = ["src.windows_app"]

[tool.setuptools.package-dir]
"" = "."
"src" = "src"

[tool.setuptools.package-data]
"*" = ["*.png", "*.jpg", "*.svg", "*.ico", "*.ttf", "*.otf", "*.json", "*.wav"]